pyproject.toml
[tool.poetry]
name = "task-scheduler"
version = "0.1.0"
description = "Asynchronous task scheduler on a remote Linux machine"
authors = ["Sefi Antebi <sefi.antebi@gmail.com>"]
readme = "README.md"
packages = [
    { include = "tasks_scheduler" }
]

[tool.poetry.dependencies]
python = "^3.8"
asyncpg = "^0.30.0"
python-dotenv = "^1.0.0"
paramiko = "^3.5.0"
psycopg2-binary = "^2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
